refactor: Change typeorm entity directory name
entity to typeorm
